Switching back to older version
Using the new 0.52.0 release from github in Pipenv is not working because Zappa looks lambda compatible packages based on this, https://github.com/Miserlou/Zappa#how-zappa-makes-packages and the new version is not available anywhere. So switching the version back to 0.51.0. This is a hack until zappa releases their official version.